Actualiza rama dev con casos de uso implementados con springboot+nextjs

+++ DeSo/src/main/java/ddb/deso/almacenamiento/DAO/HabitacionDAO.java
@@ -4,19 +4,52 @@
  */
 package ddb.deso.almacenamiento.DAO;
 
-import ddb.deso.TipoHab;
-import ddb.deso.habitaciones.Habitacion;
+import ddb.deso.service.TipoHab;
+import ddb.deso.service.habitaciones.Habitacion;
 import java.util.List;
 
 /**
+ * Contrato de acceso a datos para la entidad {@link Habitacion}.
  *
  * @author mat
  */
 public interface HabitacionDAO {
+
+    /**
+     * Persiste una nueva habitación.
+     * @param habitacion Entidad a crear.
+     */
     void crearHabitacion(Habitacion habitacion);
+
+    /**
+     * Actualiza los datos de una habitación existente.
+     * @param habitacion Entidad con datos actualizados.
+     */
     void actualizarHabitacion(Habitacion habitacion);
+
+    /**
+     * Da de baja una habitación del inventario.
+     * @param habitacion Entidad a eliminar.
+     */
     void eliminarHabitacion(Habitacion habitacion);
+
+    /**
+     * Recupera el inventario completo de habitaciones.
+     * @return Lista de habitaciones.
+     */
     List<Habitacion> listar();
+
+    /**
+     * Filtra habitaciones por su categoría o configuración.
+     * @param tipoHabitacion Enum {@link TipoHab} con la categoría buscada.
+     * @return Lista de habitaciones coincidentes.
+     */
     List<Habitacion> listarPorTipo(TipoHab tipoHabitacion);
-    Habitacion buscarPorNumero(int numero);
+
+    /**
+     * Busca una habitación por su número identificador.
+     * @param numero Identificador numérico de la habitación.
+     * @return Entidad encontrada o {@code null}.
+     */
+    Habitacion buscarPorNumero(Long numero);
 }
